% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFunctions.R
\name{find_nearest_points}
\alias{find_nearest_points}
\title{Get the names of the N-nearest points}
\usage{
find_nearest_points(point, points, negb)
}
\arguments{
\item{point}{a specific point}

\item{points}{the other points}

\item{negb}{the number of desired nearest neighbours}
}
\value{
nearest_name - the name of the nearest points
}
\description{
Get the name of the nearest point to one specified
}
