% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearUtilities.R
\name{initMedoids}
\alias{initMedoids}
\title{Initialize medoids}
\usage{
initMedoids(X, n, init_type, exclude_ids)
}
\arguments{
\item{[ndarray}{float] X: data matrix}

\item{[int]}{n: number of medoids to be selected}

\item{[string]}{init_type: rational to be used ('uniform' = randomly selected with uniform distribution, 'kpp' = k-means++ algorithm)}

\item{[ndarray}{int] exclude_ids: blacklisted ids that shouldn't be selected}
}
\value{
[ndarray int] med_ids: indices of the medoids selected
}
\description{
Initialize NC medoids with init_type rational.
}
