% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{p2corr}
\alias{p2corr}
\title{Convert p-value to correlation coefficient}
\usage{
p2corr(p, N)
}
\arguments{
\item{p}{a p-value}

\item{N}{a number of samples}
}
\value{
r a correlation coefficient
}
\description{
This functions converts an p-value into a the corresponding correlation
coefficient, using a T distribution with the provided number of samples N
minus 2 degrees of freedom
}
\examples{
N<-100
p<-0.05
p2corr(p,N)
}
