% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principalPath.R
\name{rkm}
\alias{rkm}
\title{Minimizer}
\usage{
rkm(X, init_W, s)
}
\arguments{
\item{[ndarray}{float] X: data matrix}

\item{[ndarray}{float] init_W: initial waypoints matrix}

\item{[float]}{s: regularization parameter}
}
\value{
[ndarray float] W: final waypoints matrix

[ndarray int] labels: final
}
\description{
Regularized K-mens for principal path, MINIMIZER.
}
