% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principalPath.R
\name{rkm_MS_pathvar}
\alias{rkm_MS_pathvar}
\title{Model Selection}
\usage{
rkm_MS_pathvar(models, s_span, X)
}
\arguments{
\item{s_span}{array with values of the reg parameter for each model (sorted in decreasing order, with 0 as last value)}

\item{X}{data matrix}

\item{modelsmatrix}{with path models, shape, N_models x N x (NC+2)}
}
\description{
Regularized K-menans for principal path, MODEL SELECTION, variance on waypoints interdistance.
}
