% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principalPath.R
\name{rkm_prefilter}
\alias{rkm_prefilter}
\title{Prefiltering}
\usage{
rkm_prefilter(X, boundary_ids, Nf = 200, k = 5, p = 1000, T = 0.1)
}
\arguments{
\item{[ndarray}{float] X: data matrix}

\item{[ndarray}{int] boundary_ids: start/end waypoints as sample indices}

\item{[int]}{Nf: number of filter centroids}

\item{[int]}{k: number of nearest neighbor for the penalized graph}

\item{[float]}{p: penalty factor for the penalized graph}

\item{[float]}{T: filter threshold}
}
\value{
[ndarray float] X_filtered

[ndarray int] X_labels_filtered

[ndarray int] boundary_ids_filtered

[ndarray float] X_garbage

[ndarray int] X_labels_garbage
}
\description{
Regularized K-mens for principal path, PREFILTER.
}
