% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spathialFunctions.R
\name{spathial}
\alias{spathial}
\title{Compute principal path}
\usage{
spathial(X, boundaries, NC, prefiltering)
}
\arguments{
\item{[ndarray}{float] X: data points}

\item{[ndarray}{int] boundaries: starting and ending points}

\item{[int]}{NC: number of waypoints}

\item{[ndarray}{logic] prefiltering: TRUE/FALSE}
}
\value{
[ndarray float] ppath: waypoints
}
\description{
Get the coordinates of the waypoints of the principal path
}
