% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleFunctions.R
\name{spathial_2D}
\alias{spathial_2D}
\title{2D spathial}
\usage{
spathial_2D(points)
}
\arguments{
\item{[ndarray}{float] ppath: waypoints}
}
\value{
[ndarray float] 2D_ppath: 2D coordinates of the waypoints
}
\description{
Get the 2D coordinates of each waypoint (using t-SNE algorithm for the dimensionality reduction)
}
