% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFunctions.R
\name{spathial_2D_plot}
\alias{spathial_2D_plot}
\title{2D spathial}
\usage{
spathial_2D_plot(X, X_labels, boundary_ids, spathial_res, perplexity_value)
}
\arguments{
\item{X}{data points}

\item{X_labels}{labels of the data points}

\item{ppath}{waypoints}
}
\description{
Get the 2D coordinates of each waypoint (using t-SNE algorithm for the dimensionality reduction)
}
