% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFunctions.R
\name{spathial_boundary_ids}
\alias{spathial_boundary_ids}
\title{Select starting and ending points}
\usage{
spathial_boundary_ids(X, X_labels, mode, from = NULL, to = NULL)
}
\arguments{
\item{X}{data points}

\item{X_labels}{labels of the data points}

\item{mode}{strategy for boundary selection
\itemize{
  \item 1 - centroids
  \item 2 - selected by the user
  \item 3 - insert the row name of the starting and ending points
}}

\item{from}{- starting class or row name of the starting point}

\item{to}{- ending class or row name of the ending point}
}
\value{
list
\itemize{
  \item boundary ids - The indexes of the boundaries
  \item X - The new data matrix with the boundary
  \item X_labels - The new labels of the data matrix with the boundary labels
}
}
\description{
Get the coordinates of the starting and ending points
}
