% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFunctions.R
\name{spathial_corr}
\alias{spathial_corr}
\title{Correlation}
\usage{
spathial_corr(spathial_res)
}
\arguments{
\item{ppath}{waypoints}
}
\value{
corr - correlation along the path
}
\description{
Get how much the features correlate with the path
}
