% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spathialFunctions.R
\name{spathial_way}
\alias{spathial_way}
\title{Compute principal path}
\usage{
spathial_way(X, boundaries, NC, prefiltering)
}
\arguments{
\item{X}{data points}

\item{boundaries}{starting and ending points}

\item{NC}{number of waypoints}

\item{prefiltering}{TRUE/FALSE}
}
\value{
spathial - waypoints
}
\description{
Get the coordinates of the waypoints of the principal path
}
