% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFunctions.R
\name{spathial_way_multiple}
\alias{spathial_way_multiple}
\title{Compute Principal Path}
\usage{
spathial_way_multiple(X, X_labels, boundary_ids, NC, prefiltering,
  negb = NULL)
}
\arguments{
\item{X}{data points}

\item{X_labels}{labels of the data points}

\item{boundary_ids}{starting and ending points}

\item{NC}{number of waypoints}

\item{prefiltering}{a boolean}

\item{negb}{the number of desired nearest neighbours}
}
\value{
ppath - spathial waypoints
}
\description{
Get the coordinates of the waypoints of the principal path
}
