% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFunctions.R
\name{spathialLabels}
\alias{spathialLabels}
\title{Find labels}
\usage{
spathialLabels(X, X_labels, spathial_res)
}
\arguments{
\item{X}{data points}

\item{X_labels}{labels of the data points}

\item{spathial_res}{A list of objects
\itemize{
  \item ppath: principal path from the starting point to the ending point
  \item ppath_perturbed: all the perturbed paths
}}
}
\value{
ppath_labels - labels of the waypoints
}
\description{
Get the label of each waypoint accordin to the neighbourhood
}
