% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFunctions.R
\name{spathialPlot2D}
\alias{spathialPlot2D}
\title{2D spathial}
\usage{
spathialPlot2D(X, X_labels, boundary_ids, spathial_res,
  perplexity_value = NULL, X_garbage = NULL, X_labels_garbage = NULL)
}
\arguments{
\item{X}{data points}

\item{X_labels}{labels of the data points}

\item{boundary_ids}{waypoints}

\item{spathial_res}{A list of objects
\itemize{
  \item ppath: principal path from the starting point to the ending point
  \item ppath_perturbed: all the perturbed paths
}}

\item{perplexity_value}{the value for TSNE perplexity (default is nrsamples*3/50)}

\item{X_garbage}{the data points removed during the filtering (if the prefiltering is done)}

\item{X_labels_garbage}{the labels of the data points removed during the filtering (if the prefiltering is done)}
}
\description{
Get the 2D coordinates of each waypoint (using t-SNE algorithm for the dimensionality reduction)
}
