% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFunctions.R
\name{spathialPrefiltering}
\alias{spathialPrefiltering}
\title{Prefilter data}
\usage{
spathialPrefiltering(X, X_labels, boundary_ids)
}
\arguments{
\item{X}{a data matrix}

\item{X_labels}{labels of the data points}

\item{boundary_ids}{names of the start and ending points, to be treated separately}
}
\value{
A list of objects
\itemize{
  \item X_filtered: the filtered data matrix
  \item X_labels_filtered: the filtered labels of the data points
  \item boundary_ids_filtered: the filtered boundary ids
  \item X_garbage: the data matrix that was discarded
  \item X_garbage_filtered: the filterad labels that was discarded
}
}
\description{
Regularized K-means for principal path: prefiltering
}
