% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearUtilities.R
\name{initMedoids}
\alias{initMedoids}
\title{intMedoids}
\usage{
initMedoids(X, n, init_type, boundary_ids)
}
\arguments{
\item{X}{a data matrix}

\item{n}{integer number of medoids to be selected}

\item{init_type}{character, the rational to be used
('uniform' = randomly selected with uniform distribution, 'kpp' = k-means++ algorithm)}

\item{boundary_ids}{blacklisted ids that shouldn't be selected}
}
\value{
A vevtor of medoids selected
}
\description{
Initialize NC medoids with init_type rational.
}
