% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principalPath.R
\name{rkm}
\alias{rkm}
\title{rkm}
\usage{
rkm(X, init_W, s)
}
\arguments{
\item{X}{a data matrix}

\item{init_W}{initial waypoints matrix}

\item{s}{regularization parameter}

\item{plot_ax}{boolean, whether to plot a graph}
}
\value{
W - final waypoints matrix
}
\description{
Regularized K-means for principal path, MINIMIZER.
}
\references{
'Finding Prinicpal Paths in Data Space', M.J.Ferrarotti, W.Rocchia, S.Decherchi, IEEE transactions on neural networks and learning systems 2018
}
