% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principalPath.R
\name{rkm_prefilter}
\alias{rkm_prefilter}
\title{rkm_prefilter}
\usage{
rkm_prefilter(X, boundary_ids, Nf = 200, k = 5, p = 1000, T = 0.1,
  plot_ax = FALSE)
}
\arguments{
\item{X}{a data matrix}

\item{boundary_ids}{names of the start and ending points, to be treated
separately}

\item{plot_ax}{boolean, whether the post-filtering graph should be plotted}
}
\value{
A list with letters and numbers.
\itemize{
  \item X_filtered - The filtered data matrix
  \item boundary_ids_filtered - The filtered boundary ids
  \item X_garbage - The data matrix that was discarded
}
}
\description{
Regularized K-means for principal path: prefiltering
}
