% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFunctions.R
\name{spathial_labels}
\alias{spathial_labels}
\title{Find labels}
\usage{
spathial_labels(X, X_labels, ppath)
}
\arguments{
\item{[ndarray}{float] X: data points}

\item{[ndarray}{int] X_labels: labels of the data points}

\item{[ndarray}{float] ppath: waypoints}
}
\value{
[ndarray int] ppath_labels: labels of the waypoints
}
\description{
Get the label of each waypoint accordin to the neighbourhood
}
