% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFunctions.R
\name{spathial_way}
\alias{spathial_way}
\title{Principal Path core}
\usage{
spathial_way(X, boundary_ids, NC, prefiltering)
}
\arguments{
\item{X}{data points}

\item{boundary_ids}{starting and ending points}

\item{NC}{number of waypoints}

\item{prefiltering}{a boolean}
}
\value{
ppath - spathial waypoints
}
\description{
Get the coordinates of the waypoints of the principal path
}
