% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFunctions.R
\name{spathialLabels}
\alias{spathialLabels}
\title{Find labels}
\usage{
spathialLabels(X, X_labels, spathial_res)
}
\arguments{
\item{X}{data points}

\item{X_labels}{labels of the data points}

\item{spathial_res:}{principal path from the starting point to the ending point}
}
\value{
ppath_labels: labels of the waypoints
}
\description{
Get the label of each waypoint accordin to the neighbourhood
}
\examples{
#EXAMPLE 1
# Load data matrix X
load(system.file('extdata','X.rda',package='spathial',mustWork=TRUE))
# Load description vector X_labels
load(system.file('extdata','X_labels.rda',package='spathial',mustWork=TRUE))
# Run spathialBoundary
boundaryRes <- spathialBoundaryIds(X, X_labels, mode=2, from=3, to=6)
#Set the number of waypoints
NC <- 20
# Run spathialWay without prefiltering
spathial_res <- spathialWay(boundaryRes$X, boundaryRes$X_labels, boundaryRes$boundary_ids, NC)
#Run spathialLabels with spathial_res
labels <- spathialLabels(boundaryRes$X, boundaryRes$X_labels, spathial_res)
labels

#EXAMPLE 2
# Load data matrix X
load(system.file('extdata','X.rda',package='spathial',mustWork=TRUE))
# Load description vector X_labels
load(system.file('extdata','X_labels.rda',package='spathial',mustWork=TRUE))
# Run spathialBoundary
boundaryRes <- spathialBoundaryIds(X, X_labels, mode=2, from=3, to=6)
# Run spathial spathialPrefilterinh with the output of the function spathialBoundaryIds
filterRes <- spathialPrefiltering(boundaryRes$X, boundaryRes$X_labels, boundaryRes$boundary_ids)
#Set the number of waypoints
NC <- 20
# Run spathialWay with the filtered data
spathial_res <- spathialWay(boundaryRes$X[filterRes$mask,], boundaryRes$X_labels[filterRes$mask], filterRes$boundary_ids, NC)
#Run spathialLabels with spathial_res
labels <- spathialLabels(boundaryRes$X[filterRes$mask,], boundaryRes$X_labels[filterRes$mask], spathial_res)
labels
}
