% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFunctions.R
\name{spathialStatistics}
\alias{spathialStatistics}
\title{Correlation}
\usage{
spathialStatistics(spathial_res)
}
\arguments{
\item{spathial_res}{A list of objects
\itemize{
  \item ppath: principal path from the starting point to the ending point
  \item ppath_perturbed: all the perturbed paths
}}
}
\value{
A list of objects
\itemize{
  \item ppath: spathial waypoints
  \item ppath_parturbed: all the spathial waypoints for each perturbation
}
}
\description{
Get how much the features correlate with the path
}
