% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFunctions.R
\name{spathialStatistics}
\alias{spathialStatistics}
\title{Correlation}
\usage{
spathialStatistics(spathial_res)
}
\arguments{
\item{spathial_res}{A list of objects
\itemize{
  \item ppath: principal path from the starting point to the ending point
  \item ppath_perturbed: all the perturbed paths
}}
}
\value{
A list of objects
\itemize{
  \item ppath: spathial waypoints
  \item ppath_parturbed: all the spathial waypoints for each perturbation
}
}
\description{
Get how much the features correlate with the path
}
\examples{
load("X.rda")
load("X_labels.rda")
boundaryRes <- spathialBoundaryIds(X, X_labels, mode=2, from=3, to=6)
NC <- 20
spathialRes <- spathialWay(boundaryRes$X, boundaryRes$X_labels, boundaryRes$boundary_ids, NC, neighbors = 1)
statistics <- spathialStatistics(spathialRes)
}
