% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplelFunctions.R, R/runFunctions.R
\name{spathialWay}
\alias{spathialWay}
\title{spathialWay}
\usage{
spathialWay(X, boundary_ids, NC, prefiltering)

spathialWay(X, boundary_ids, NC, prefiltering)
}
\arguments{
\item{X}{data points}

\item{boundary_ids}{starting and ending points}

\item{NC}{number of waypoints}

\item{prefiltering}{a boolean}

\item{X}{data points}

\item{boundary_ids}{starting and ending points}

\item{NC}{number of waypoints}

\item{prefiltering}{a boolean}
}
\value{
spathial waypoints

spathial waypoints
}
\description{
Get the coordinates of the waypoints of the principal path

Get the coordinates of the waypoints of the principal path
}
