% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFunctions.R
\name{spathialWay}
\alias{spathialWay}
\title{Compute Principal Path}
\usage{
spathialWay(X, X_labels, boundary_ids, NC = 50, neighbors = NULL)
}
\arguments{
\item{X}{data points}

\item{X_labels}{labels of the data points}

\item{boundary_ids}{starting and ending points}

\item{NC}{number of waypoints}

\item{neighbors}{the number of desired nearest neighbours}
}
\value{
A list of objects
\itemize{
  \item ppath: spathial waypoints
  \item ppath_parturbed: all the spathial waypoints for each perturbation
}
}
\description{
Get the coordinates of the waypoints of the principal path
}
