% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleFunctions.R
\name{spathialWay}
\alias{spathialWay}
\title{Compute Principal Path}
\usage{
spathialWay(X, boundary_ids, NC, prefiltering)
}
\arguments{
\item{X}{data points}

\item{NC}{number of waypoints}

\item{prefiltering}{a boolean}

\item{boundaries}{starting and ending points}
}
\value{
spathial waypoints
}
\description{
Get the coordinates of the waypoints of the principal path
}
