% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFunctions.R
\name{spathialWay}
\alias{spathialWay}
\title{Compute Principal Path}
\usage{
spathialWay(X, X_labels, boundary_ids, NC = 50, neighbors = 0)
}
\arguments{
\item{X}{data points}

\item{X_labels}{labels of the data points}

\item{boundary_ids}{starting and ending points}

\item{NC}{number of waypoints}

\item{neighbors}{the number of desired nearest neighbours}
}
\value{
A list of objects
\itemize{
  \item ppath: spathial waypoints
  \item ppath_parturbed: all the spathial waypoints for each perturbation
}
}
\description{
Get the coordinates of the waypoints of the principal path
}
\examples{
#EXAMPLE 1
# Load data matrix X
load(system.file('extdata','X.rda',package='spathial',mustWork=TRUE))
# Load description vector X_labels
load(system.file('extdata','X_labels.rda',package='spathial',mustWork=TRUE))
# Run spathialBoundary
boundaryRes <- spathialBoundaryIds(X, X_labels, mode=2, from=3, to=6)
#Set the number of waypoints
NC <- 20
# Run spathialWay without prefilterint
spathialRes <- spathialWay(boundaryRes$X, boundaryRes$X_labels, boundaryRes$boundary_ids, NC, neighbors = 1)
spathialRes

#EXAMPLE 2
# Load data matrix X
load(system.file('extdata','X.rda',package='spathial',mustWork=TRUE))
# Load description vector X_labels
load(system.file('extdata','X_labels.rda',package='spathial',mustWork=TRUE))
# Run spathialBoundary
boundaryRes <- spathialBoundaryIds(X, X_labels, mode=2, from=3, to=6)
# Run spathial spathialPrefilterinh with the output of the function spathialBoundaryIds
filterRes <- spathialPrefiltering(boundaryRes$X, boundaryRes$X_labels, boundaryRes$boundary_ids)
#Set the number of waypoints
NC <- 20
# Run spathialWay with the filtered data
spathialRes <- spathialWay(boundaryRes$X[filterRes$mask,], boundaryRes$X_labels[filterRes$mask], filterRes$boundary_ids, NC, neighbors = 1)
spathialRes
}
